Load ROM instead of starting with BIOS
I realized that a Game Boy without a cartridge isn't worth anything:
the BIOS starts up until it wants to read the game logo data from
memory, and then it hangs indefinitely.

To avoid implementing the "memory shadowing" that the GB does by
effectively mapping the BIOS over the first 0x100 bytes of memory,
let's just start with a ROM only for now.